Refactor BookList, the books variable that holds and array of predefined books its imported from ./src/components/books/books
